Move syncSelectionScopeList to IModelLoader
